refactor(frontend/src/features/commands
/create/CreateForm.jsx):

move CreateForm.jsx to features/command/create

for modular structure
